+++ x-pack/auditbeat/module/system/package/package.go
@@ -236,7 +236,7 @@ func New(base mb.BaseMetricSet) (mb.MetricSet, error) {
 	}
 
 	if config.PackageSuidDrop != nil {
-		if os.Getuid() != 0 && int(*config.PackageSuidDrop) != os.Getuid() {
+		if os.Getuid() != 0 && int(*ms.config.PackageSuidDrop) != os.Getuid() {
 			return nil, fmt.Errorf("package.rpm_drop_to_suid is set to %d, but we're running as a different non-root user", config.PackageSuidDrop)
 		}
 		ms.log.Debugf("Dropping to EUID %d for RPM API calls", *ms.config.PackageSuidDrop)
