+++ x-pack/auditbeat/module/system/package/package_linux_test.go
@@ -9,6 +9,7 @@ package pkg
 import (
 	"os"
 	"strconv"
+	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -17,11 +18,63 @@ import (
 	"github.com/elastic/elastic-agent-libs/logp"
 )
 
+func TestRPMParallel(t *testing.T) {
+	currentUID := os.Getuid()
+	if currentUID != 0 {
+		t.Skipf("can only run as root")
+	}
+	logp.DevelopmentSetup()
+
+	count := 20
+	waiter := sync.WaitGroup{}
+	waiter.Add(count)
+
+	useUID := getUser(t)
+
+	t.Logf("Starting...")
+	for i := 0; i < count; i++ {
+		inner := i
+		go func() {
+			defer waiter.Done()
+			testMs := MetricSet{
+				log: logp.L(),
+				config: config{
+					PackageSuidDrop: &useUID,
+				},
+			}
+
+			pkgList, err := testMs.getPackages()
+			require.NoError(t, err)
+
+			t.Logf("got %d packages from %d", len(pkgList), inner)
+		}()
+
+	}
+	waiter.Wait()
+}
+
 func TestWithSuid(t *testing.T) {
 	currentUID := os.Getuid()
 	if currentUID != 0 {
 		t.Skipf("can only run as root")
 	}
+
+	useUID := getUser(t)
+	testMs := MetricSet{
+		log: logp.L(),
+		config: config{
+			PackageSuidDrop: &useUID,
+		},
+	}
+
+	packages, err := testMs.getPackages()
+	require.NoError(t, err)
+
+	require.NotZero(t, packages)
+	t.Logf("got %d packages", len(packages))
+}
+
+func getUser(t *testing.T) int64 {
 	// pick a user to drop to
 	userList, err := user.GetUsers(false)
 	require.NoError(t, err)
@@ -35,17 +88,5 @@ func TestWithSuid(t *testing.T) {
 			break
 		}
 	}
-
-	testMs := MetricSet{
-		log: logp.L(),
-		config: config{
-			PackageSuidDrop: &useUID,
-		},
-	}
-
-	packages, err := testMs.getPackages()
-	require.NoError(t, err)
-
-	require.NotZero(t, packages)
-	t.Logf("got %d packages", len(packages))
+	return useUID
 }
